creted a ReleasenetForksModule, which is basically a copy of StokenetForksModule
