+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/statecomputer/forks/modules/ReleasenetForksModule.java
@@ -62,56 +62,98 @@
  * permissions under this License.
  */
 
-package com.radixdlt.statecomputer.forks.testing;
+package com.radixdlt.statecomputer.forks.modules;
 
-import com.google.common.collect.ImmutableSet;
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.ProvidesIntoSet;
 import com.radixdlt.application.system.FeeTable;
 import com.radixdlt.application.tokens.Amount;
-import com.radixdlt.statecomputer.forks.CandidateForkConfig;
+import com.radixdlt.application.tokens.state.PreparedStake;
+import com.radixdlt.application.tokens.state.PreparedUnstakeOwnership;
+import com.radixdlt.application.tokens.state.TokenResource;
+import com.radixdlt.application.validators.state.AllowDelegationFlag;
+import com.radixdlt.application.validators.state.ValidatorFeeCopy;
+import com.radixdlt.application.validators.state.ValidatorMetaData;
+import com.radixdlt.application.validators.state.ValidatorOwnerCopy;
+import com.radixdlt.application.validators.state.ValidatorRegisteredCopy;
 import com.radixdlt.statecomputer.forks.ForkBuilder;
 import com.radixdlt.statecomputer.forks.RERulesConfig;
 import com.radixdlt.statecomputer.forks.RERulesVersion;
-import java.util.Collections;
 
-public final class TestingForksModuleV5 extends AbstractModule {
-  @ProvidesIntoSet
-  ForkBuilder genesis() {
-    return new ForkBuilder(
-        "genesis", 0L, RERulesVersion.OLYMPIA_V1, RERulesConfig.testingDefault(500));
-  }
+import java.util.Map;
+import java.util.OptionalInt;
+import java.util.Set;
+import java.util.regex.Pattern;
 
-  @ProvidesIntoSet
-  ForkBuilder fork1() {
-    return new ForkBuilder(
-        "fork-1", 2L, RERulesVersion.OLYMPIA_V1, RERulesConfig.testingDefault(500));
-  }
+import static com.radixdlt.constraintmachine.REInstruction.REMicroOp.MSG;
+
+public final class ReleasenetForksModule extends AbstractModule {
+  private static final Set<String> RESERVED_SYMBOLS =
+      Set.of("xrd", "xrds", "exrd", "exrds", "rad", "rads", "rdx", "rdxs", "radix");
 
   @ProvidesIntoSet
-  ForkBuilder fork2() {
+  ForkBuilder releasenetGenesis() {
     return new ForkBuilder(
-        "fork-2", 7L, RERulesVersion.OLYMPIA_V1, RERulesConfig.testingDefault(500));
-  }
-
-  @ProvidesIntoSet
-  ForkBuilder fork4() {
-    return new ForkBuilder(
-        "fork-4",
-        16L,
+        "stokenet-genesis",
+        0L,
         RERulesVersion.OLYMPIA_V1,
-        RERulesConfig.testingDefault(
-            500, FeeTable.create(Amount.ofMicroTokens(100), Collections.emptyMap())));
+        new RERulesConfig(
+            RESERVED_SYMBOLS,
+            Pattern.compile("[a-z0-9]+"),
+            FeeTable.create(
+                Amount.ofMicroTokens(200), // 0.0002XRD per byte fee
+                Map.of(
+                    TokenResource.class, Amount.ofTokens(100), // 100XRD per resource
+                    ValidatorRegisteredCopy.class, Amount.ofTokens(5), // 5XRD per validator update
+                    ValidatorFeeCopy.class, Amount.ofTokens(5), // 5XRD per register update
+                    ValidatorOwnerCopy.class, Amount.ofTokens(5), // 5XRD per register update
+                    ValidatorMetaData.class, Amount.ofTokens(5), // 5XRD per register update
+                    AllowDelegationFlag.class, Amount.ofTokens(5), // 5XRD per register update
+                    PreparedStake.class, Amount.ofMilliTokens(500), // 0.5XRD per stake
+                    PreparedUnstakeOwnership.class, Amount.ofMilliTokens(500) // 0.5XRD per unstake
+                    )),
+            (long) 1024 * 1024, // 1MB max user transaction size
+            OptionalInt.of(50), // 50 Txns per round
+            10_000, // Rounds per epoch
+            500, // Two weeks worth of epochs
+            Amount.ofTokens(90), // Minimum stake
+            500, // Two weeks worth of epochs
+            Amount.ofMicroTokens(2307700), // Rewards per proposal
+            9800, // 98.00% threshold for completed proposals to get any rewards,
+            100, // 100 max validators
+            MSG.maxLength()));
   }
 
   @ProvidesIntoSet
-  ForkBuilder fork5() {
+  ForkBuilder releasenetV2() {
     return new ForkBuilder(
-        "fork-5",
-        ImmutableSet.of(new CandidateForkConfig.Threshold((short) 75 /* 75% */, 1)),
-        24,
-        Long.MAX_VALUE,
+        "stokenet-v2",
+        7914L,
         RERulesVersion.OLYMPIA_V1,
-        RERulesConfig.testingDefault(500));
+        new RERulesConfig(
+            RESERVED_SYMBOLS,
+            Pattern.compile("[a-z0-9]+"),
+            FeeTable.create(
+                Amount.ofMicroTokens(190), // 0.00019XRD per byte fee
+                Map.of(
+                    TokenResource.class, Amount.ofTokens(100), // 100XRD per resource
+                    ValidatorRegisteredCopy.class, Amount.ofTokens(5), // 5XRD per validator update
+                    ValidatorFeeCopy.class, Amount.ofTokens(5), // 5XRD per register update
+                    ValidatorOwnerCopy.class, Amount.ofTokens(5), // 5XRD per register update
+                    ValidatorMetaData.class, Amount.ofTokens(5), // 5XRD per register update
+                    AllowDelegationFlag.class, Amount.ofTokens(5), // 5XRD per register update
+                    PreparedStake.class, Amount.ofMilliTokens(500), // 0.5XRD per stake
+                    PreparedUnstakeOwnership.class, Amount.ofMilliTokens(500) // 0.5XRD per unstake
+                    )),
+            (long) 1024 * 1024, // 1MB max user transaction size
+            OptionalInt.of(50), // 50 Txns per round
+            10_000, // Rounds per epoch
+            500, // Two weeks worth of epochs
+            Amount.ofTokens(90), // Minimum stake
+            500, // Two weeks worth of epochs
+            Amount.ofMicroTokens(2307700), // Rewards per proposal
+            9800, // 98.00% threshold for completed proposals to get any rewards,
+            100, // 100 max validators
+            MSG.maxLength()));
   }
 }
