+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/statecomputer/forks/modules/testing/TestingForksModuleV1.java
@@ -62,17 +62,15 @@
  * permissions under this License.
  */
 
-package com.radixdlt.statecomputer.forks.testing;
+package com.radixdlt.statecomputer.forks.modules.testing;
 
-import com.google.common.collect.ImmutableSet;
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.ProvidesIntoSet;
-import com.radixdlt.statecomputer.forks.CandidateForkConfig;
 import com.radixdlt.statecomputer.forks.ForkBuilder;
 import com.radixdlt.statecomputer.forks.RERulesConfig;
 import com.radixdlt.statecomputer.forks.RERulesVersion;
 
-public final class TestingForksModuleV2 extends AbstractModule {
+public final class TestingForksModuleV1 extends AbstractModule {
   @ProvidesIntoSet
   ForkBuilder genesis() {
     return new ForkBuilder(
@@ -83,16 +81,5 @@ public final class TestingForksModuleV2 extends AbstractModule {
   ForkBuilder fork1() {
     return new ForkBuilder(
         "fork-1", 2L, RERulesVersion.OLYMPIA_V1, RERulesConfig.testingDefault(500));
-  }
-
-  @ProvidesIntoSet
-  ForkBuilder fork2() {
-    return new ForkBuilder(
-        "fork-2",
-        ImmutableSet.of(new CandidateForkConfig.Threshold((short) 6000 /* 60% */, 1)),
-        7L,
-        Long.MAX_VALUE,
-        RERulesVersion.OLYMPIA_V1,
-        RERulesConfig.testingDefault(500));
   }
 }
