+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/statecomputer/forks/modules/testing/TestingForksModuleV2.java
@@ -62,20 +62,17 @@
  * permissions under this License.
  */
 
-package com.radixdlt.statecomputer.forks.testing;
+package com.radixdlt.statecomputer.forks.modules.testing;
 
 import com.google.common.collect.ImmutableSet;
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.ProvidesIntoSet;
-import com.radixdlt.application.system.FeeTable;
-import com.radixdlt.application.tokens.Amount;
 import com.radixdlt.statecomputer.forks.CandidateForkConfig;
 import com.radixdlt.statecomputer.forks.ForkBuilder;
 import com.radixdlt.statecomputer.forks.RERulesConfig;
 import com.radixdlt.statecomputer.forks.RERulesVersion;
-import java.util.Collections;
 
-public final class TestingForksModuleV4 extends AbstractModule {
+public final class TestingForksModuleV2 extends AbstractModule {
   @ProvidesIntoSet
   ForkBuilder genesis() {
     return new ForkBuilder(
@@ -91,20 +88,11 @@ public final class TestingForksModuleV4 extends AbstractModule {
   @ProvidesIntoSet
   ForkBuilder fork2() {
     return new ForkBuilder(
-        "fork-2", 7L, RERulesVersion.OLYMPIA_V1, RERulesConfig.testingDefault(500));
-  }
-
-  @ProvidesIntoSet
-  ForkBuilder fork4() {
-    return new ForkBuilder(
-        "fork-4",
-        ImmutableSet.of(
-            new CandidateForkConfig.Threshold((short) 7500 /* 75% */, 3),
-            new CandidateForkConfig.Threshold((short) 9000 /* 90% */, 1)),
-        11L,
+        "fork-2",
+        ImmutableSet.of(new CandidateForkConfig.Threshold((short) 6000 /* 60% */, 1)),
+        7L,
         Long.MAX_VALUE,
         RERulesVersion.OLYMPIA_V1,
-        RERulesConfig.testingDefault(
-            500, FeeTable.create(Amount.ofMicroTokens(100), Collections.emptyMap())));
+        RERulesConfig.testingDefault(500));
   }
 }
