+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/statecomputer/forks/modules/testing/TestingForksModuleV3.java
@@ -62,15 +62,20 @@
  * permissions under this License.
  */
 
-package com.radixdlt.statecomputer.forks.testing;
+package com.radixdlt.statecomputer.forks.modules.testing;
 
+import com.google.common.collect.ImmutableSet;
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.ProvidesIntoSet;
+import com.radixdlt.application.system.FeeTable;
+import com.radixdlt.application.tokens.Amount;
+import com.radixdlt.statecomputer.forks.CandidateForkConfig;
 import com.radixdlt.statecomputer.forks.ForkBuilder;
 import com.radixdlt.statecomputer.forks.RERulesConfig;
 import com.radixdlt.statecomputer.forks.RERulesVersion;
+import java.util.Collections;
 
-public final class TestingForksModuleV1 extends AbstractModule {
+public final class TestingForksModuleV3 extends AbstractModule {
   @ProvidesIntoSet
   ForkBuilder genesis() {
     return new ForkBuilder(
@@ -81,5 +86,23 @@ public final class TestingForksModuleV1 extends AbstractModule {
   ForkBuilder fork1() {
     return new ForkBuilder(
         "fork-1", 2L, RERulesVersion.OLYMPIA_V1, RERulesConfig.testingDefault(500));
+  }
+
+  @ProvidesIntoSet
+  ForkBuilder fork2() {
+    return new ForkBuilder(
+        "fork-2", 7L, RERulesVersion.OLYMPIA_V1, RERulesConfig.testingDefault(500));
+  }
+
+  @ProvidesIntoSet
+  ForkBuilder fork3() {
+    return new ForkBuilder(
+        "fork-3",
+        ImmutableSet.of(new CandidateForkConfig.Threshold((short) 8000 /* 80% */, 1)),
+        9L,
+        10L,
+        RERulesVersion.OLYMPIA_V1,
+        RERulesConfig.testingDefault(
+            500, FeeTable.create(Amount.ofMicroTokens(200), Collections.emptyMap())));
   }
 }
